New implementation of edit_last() to enable longer loglines and thus make room for record of QRG
needs a New mechanism to get and store editbuffer

Former use of logline_edit field, which was shared with some other
code for displaying the last 5 qsos, was error prone.
Use a separate editbuffer instead and get the content from the qsos[]
database every time we start to edit a line.
Write content of editbuffer back to qsos[] and file after finishing
the edit.

Furthermore refactor out common code

- highlight and unhighlite line
- limit len of displayed output line
- make number of lines easier to configure
- correct handling of empty log file
- fix edit of upper line
!! Be aware that edit on ';' notes is broken
- better handling of <Tab> to hop from field to field
- move definition of logline_edit fields into main.c