Correct GPL boilerplate in sources
The license boilerplate in the source files incorrectly referenced the
Library General Public License.  Given that only 'COPYING' (GPL 2.0) is
included in the tlf directory, and all other documentation references
the GPL, the reference to 'Library' is presumed to be incorrect.   With
the concurrance of Tom and Rein, this patch set removes the word
'Library' from the boilerplate in all .c and .h files where it occured.

Include minor Free Software Foundation mailing address correction with
this patch set.