Move onechar() routine and make it static
We hide the onechar() routine to force the use of the new
key_get() and key_poll() routines.

That is a first step to switch to a different handling of key input.
The next step will be a switch to curses keypad mode where curses
decodes the ESC sequences itself. That will allow for:
- handling more terminal types
- resizing the program window
- accepting mouse input