extend zonescore[] array to full number of bands
Until now there was a distinction between counting things for all bands
(WARC and non WARC) and counting zones and countries for contest bands
only.
The handling can be simplified if we unify the handling and count
zones and countries for ALL bands. We can then ignore the non WARC
entries for contest scoring.
Furthermore it is more easy to add additionla bands later.