Switch Autoconf Archive macros for Curses search
Use the AX_WITH_CURSES and AX_WITH_CURSES_PANEL macros to determine the
library and header files location for curses.  For now the
AX_WITH_CURSES macro is edited to only look for ncurses or curses and
not ncursesw.  Also the configure options have been commented out as
ncurses is required.

Macros are set in config.h to allow testing for the proper header file
location to include.