Load ncurses.h via tlf_curses.h
Regarding Tlf Issue #51, this patch provides consistency in the form of
having the build system determine which ncurses header file is
available.

The build system will set relevant macros in config.h and boilerplate in
tlf_curses.h will select the proper include path based on the tests.
This patch confines the boiler plate to one file resulting in minimal
editing of the other source files.