Use AX_CFLAGS_WARN_ALL to determine compiler warnings
As CFLAGS is a a precious user variable, use the AX_CFLAGS_WARN_ALL
macro and have it assign the warning flag(s) to AM_CFLAGS which is then
substituted into the Makefile.  This macro determines the best warning
for various compilers.

This macro depends on the AX_APPEND_FLAG and AX_REQUIRE_DEFINED macros
(in separate files).