Define PACKAGE_DATA_DIR in src/Makefile.am
Automake provides the predefined 'pkgdatadir' Makefile variable which is
the same path as the previously caclulated PACKAGE_DATA_DIR value in
configure.ac.  Use it to set PACKAGE_DATA_DIR as a -D value to the
preprocessor via AM_CPPFLAGS in src/Makefile.am which is passed to the
source files during preprocessing.  This conforms with current Autotools
convention.

As PACKAGE_DATA_DIR is longer defined in config.h, remove its include
directive from those files that were only including it for the
PACKAGE_DATA_DIR definition.