Process input keystrokes using keypad()
Ncurses provides the keypad() function that maps most escaped key codes
to named constants found in curses.h.  The onechar() function has been
reworked to minimize the amount of key escapes that need to be caught.
Those that are caught are assigned values to match the Ncurses declared
constants.  Keys were tested on the Linux console, and the following X11
terminal emulators, Xfce Terminal, Gnome Terminal, Xterm, and Qterm.

Update the tested constants to match the Ncurses defined constants in
the rest of the sort where key_get() and key_poll() are called.

Ncurses will honor the value of the environment variable ESCDELAY but to
avoid the users having to set it, use the set_escdelay() function to set
the delay to 25 mS (default is 1000 mS when unset) when the environment
variable is unset.  Should that be too short, users can set it via the
usual method, i.e. 'export ESCDELAY=125' would set a delay of 125 mS.