Initialize dupe value when a new spot is created
Checking with Valgrind memcheck generated in the report that,
"Conditional jump or move depends on uninitialised value(s)" in
bmdata_write_file (bandmap.c:134).  The result is that various values
could be written for the dupe value in .bmdata.dat.  Fixed by setting
the dupe variable to 0 in bandmap_addspot().  This variable is only used
in bandmap_show() and is never stored back into the allspots GList.