refactor handling of send cw or digi data
Cleanup code so that all data for cw or digi keying goes through the
write_keyer() interface.
That means anyone who wants to send code has to use the keyer_append()
function which (in an thread safe manner) appends the data to a
keybuffer.
The actual keying is done by a background thread which regularly has to
call the write_keyer() function.