Migrate packet screen to ncurses panels
Old implementation used a separate newterm call to get a separate
display for packet screen -- separate from normal logger display.
It was quite a hack but worked ok.
In new ncurses6 newterm does no longer give us a separate screen. So we
had to switch to use ncurses panels for the packet screen.
A popup panel covers inow the whole screen when we switch to packet. That
window is split into two derived windows for display of cluster messages
and a second one for input - as did the old implementation.