output of cluster message needs a packet screen
Each cluster message (external or from internal network between multiple
loggers) will be processed in several steps: putting it on cluster
screen, looking for a personal conversation, adding to bandmap.
The change guarantees that it will only be added to the cluster screen
if such an object got initialized. (That would not be the case if we
started with the 'nopacket'-option or have no cluster configured).