Fix wrong return value of callinput()
This PR fixes the "long CQ bug": when sending a longer series of
unanswered CQ's the cursor sometimes jumps to the exchange field. This
causes F1 to send your call instead of CQ, so the op has to press either
Esc or Tab to get back to CQ'ing. Quite annoying, but keeps op awake.
:-)

An Easter egg version of this "feature" is:

    start tlf
    press F3 14 times
    CW keyer window pops up.

The bug is a result of two issues.

  In callinput() key codes returned as char although they fit in an int
  only. E.g. F1 is 0x109 and F3 is 0x10B. On return only the lower 8
  bits are kept and thus F1 is transformed into a Tab/Ctrl-I (0x09) and
  F3 into Ctrl-K (0x0B).

  callinput() is supposed to return to logit() only when some non-local
  keypress is to be handled (e.g. Enter).  Instead it returns when its
  internal counter i reaches 14:

    for (i = strlen(hiscall); i <= 13; i++) {

  The idea may have been an optimization to avoid excessive calls to
  strlen and track call length in i.  But this doesn't bring anything as
  this check is done only once per keypress and in the inner loop [while
  (x < 1)] strlen and other stuff is happening 100 times a second
  anyway.