Add DIGI_RIG_MODE config option
Allows to define the rig mode to use if DIGI mode is selected.

This option can be one of RTTY, RTTYR, USB, or LSB.  If set, the
rig mode is changed to it any time digi mode is activated.

The mode is also set on startup.

This commit also cleans up the commads array so that MAX_COMMANDS
doesn't need to be updated every time a new command is added.