Support CT9 cty.dat files
This adds three new overrides, lat/lon, continent, and time offset.
showinfo() now needs a prefix, so create add a getctydata_pfx()
function which returns the prefix rather than the country.

If a callsign in cty.dat starts with an '=', it will only use
exact matching, not prefix matching.  To allow backward compatability,
Tlf remembers if it saw any '=' prefixes and if so, find_full_match()
will only match those.  If there were no '=' prefixes, the behaviour
is as before, and it will match any prefix.