RTTY contests use a space, not newline at end of messages.
Convert the messages from trailing \n to trailing space when digi
mode is set, and convert them back for cw.  They are unchanged in
ssb since I don't think they're used.

It doesn't look like anything requires a \n at the end, but I'm not
sure about the various MFJ keyers.

It may be better to have separate digi messages and a macro
character for CR and/or LF, but I think that would need to wait
until after the SARTG WW RTTY contest.