Allow wider adjustment QRG of grabbed station
The reported QRG of spotted stations may not be exact. So after a grab
of it from the bandmap list you may want to adjust the QRG a little.

On the other side after a grab if you just tune away you would like the
call input field to become empty automatically.

The patch widens the allowed adjustment to +/-500 Hz. That should
hopefully be enough.

Problem reported by Fred DH5FS.