fix frequency adjustment function (Ctrl-F)
Replace key_get() with key_poll().
In order to avoid flickering the order of freq_display() and time_update()
was reversed. Minor code cleanup was done in freq_display.c.

Ctrl-F is now completely live in the sense that it also takes over freq
changes from the rig. The cluster display below Ctrl-F is also updated
and the displayed values are all in sync. For TRX there is a delay of 1
sec but that's tolerable as it's the normal behavior.