improve command line argument handling
Switched to Argp as it provides a number of features (usage, version, error messages, long arguments, ...) out of the box.

The original issue was that argument option had to be written without space (-fmyfile.dat). Now both this, the "normal' format (-f myfile.dat), and the long format (--config=myfile.dat) are also supported.