Fix CT compatible keys and document CT mode
The Insert and + keys were sending an incorrect macro leading to
opposite behavior from their function in classic CT.  Insert now sends
the RST macro and + sends the TU macro and logs the QSO.

The Insert key in the callsign field now advances the cursor to the
exchange field causing autofill to be invoked, if used for a particular
contest.

The Insert and + keys now work in the exchange field, sending their
assigned macros.

Document CT Compatible mode in the manual page.