Change rst (#172)
Allow scrolling up and down received and sent rapport between 33(9) and 59(9) if CHANGE_RST is set.

Extend keywords so that you can provide a list of most used RST values for other stations rapport which will then be used for scrolling, e.g. CHANGE_RST=339,479,599.

Allowed range is 33..39,43..49,53..59.
A value for T can be provided, but will be ignored, e.g. CHANGE_RST=338,.. will be accepted but always send as 339.