Modify soundlog recording and playback
* Allow for soundlog script in PATH

As installation places the soundlog script into ${prefix}/bin,
avoid requiring the user to copy it to the ~/tlf/soundlogs directory.

It works just fine from PATH when called from inside the directory.

* Query the environment for the value of HOME

The opendir() function is not in a context where it understands
the value of $HOME, unlike the system() calls used elsewhere which
operate in a user's shell environment.  To make this work, query the
environment for the value of HOME and concatenate it to the hard coded
directory and pass that to opendir().

Add some error handling so if in the future the directory is exposed to
user configuration, the user won't be completely in the dark if there is
a typo in the entered path.