Make SoX play command the default
The -w option on the rec commandline probably dates from the Open Sound
System (OSS) version of rec.  This option is not recognized by the SoX
version of rec which is the only version available on modern Debian.

This command line has been commented out so users can restore it easily
if needed.

Instead add an example command for Sox rec as the default.  The -c 1
option forces single channel recording.  This needs to be documented in
the manual page so users can remove it as two channel recording is the
default for this version of rec.  Two channel recording may be useful
for transceivers that output two channel audio, such as the Elecraft K3.

The -q option makes the rec utility quiet, quelling its normal output.