Add a HAMLIBKEYER option to send CW via hamlib
Most modern rigs can send CW using an internal keyer. Hamrig exposes
this functionality in the send_morse and stop_morse functions.

This introduces a new keyer type HAMLIB_KEYER and when this is set in
the config file with the HAMLIBKEYER options sends CW via hamlib.

Hamlib doesn't (yet) expose setting the speed of the CW so this will
be sent at whatever the rig's keyer is set to. This also ignores speed
change commands in the CW.

This simplifies the setup enormously meaning no external keyer is
needed.