Don't print a complaint if either Alt-PgUp/Dn or Ctrl-PgUp/Dn work
There is an annoying delay on startup if the terminal doesn't support
one of Alt-PgUp Alt-PgDn Ctrl-PgUp Ctrl-PgDn, even if the other
combination works. If fact this even happens with rxvt-unicode which the
manpage advertises as "does not eat Ctrl-PgUp".

Change logic such that the delay only happens if none of the keys work.