Test rules (#379)
Add a test mechanism tests for specific contest rules.
For each contest to be covered one has to set up a minimal logcfg+rule+log in a subdir of test/rules. The test starts in that directory TLF and expects that no changes happen to the log file. This ensures that score and mults are evaluated correctly.