Use full path for 'rec' in scripts/soundlog
Due to the pattern matching of the '-f' option to 'pkill' in the man
page example, use the full path which should avoid matching the command
line of other processes.  For references see:

https://lists.debian.org/debian-user/2023/08/msg00406.html
https://lists.debian.org/debian-user/2023/09/msg00211.html
https://lists.debian.org/debian-user/2024/02/msg00800.html

https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=1053923

In short the '-f rec' option matched the '--directory' option of
'gnome-keyring-daemon' and killed it at the top of each hour when
'pkill -f rec' was run.

Modifying the script and man page to use the full path (whatever that
might be, typically '/usr/bin/rec' in Linux but could be elsewhere)
should go along way to helping someone avoid similar pain.