Fix pkill man page (#427)
* Cleans up an ambiguous example in the man page

Using parameter -f to 'pkill' kills the first program which contains the mentioned
'rec' string in its command line. That may not be the program with the
name 'rec'.

* Replace backticks with $() in scripts/soundlog

As backticks are generally hard to see, use the synonymous $() construct
to generate the name for the audio file.  $() is preferred for capturing
the output of a command into a shell variable.

---------

Co-authored-by: Nate Bargmann <n0nb@n0nb.us>