switch to "private" xmlrpc client mode
The problem is that the xmlrpc_client_call_server_params()
or other function using the "global" client mode return
a non-initialized result pointer (a local variable containing garbage)
in case of some error situations.
This makes impossible to decide whether the result should be free'd or not.
The xmlrpc documentation contains no hint on this.
The "private" functions on the other can use a pointer passed by the caller,
so in this case if the initial NULL is untouched then it hasn't been allocated
yet, hence no free'ing is needed.