Make Enter and similar keys keep working when editing call and exchange
When editing the call field, Enter didn't anything. When editing the
exchange field, Enter at least exited the edit mode, but did not log the
QSO.

Make Enter, Backslash, Tab and similar keys exit editing, and perform
their normal function so QSOs can be logged faster.

This also fixes the bug that editing the call did not terminate on
unhandled keys (like End in that context).