Start immediately with the main screen (#497)
There is little point in delaying startup when there are no
"interesting" messages. Remove sleep(1); the user can still request a
verbose run with `tlf -v` (this is the default on the first startup
anyway).

When having to restart tlf during a contest (after changing some setting
for example), that second actually makes a difference.